[refactor] 냉장고 엔티티 제거로 fridge -> ingredient 패키지명 변경
