+++ src/main/java/refooding/api/domain/ingredient/repository/MemberIngredientCustomRepository.java
@@ -1,8 +1,8 @@
-package refooding.api.domain.fridge.repository;
+package refooding.api.domain.ingredient.repository;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
-import refooding.api.domain.fridge.dto.response.IngredientResponse;
+import refooding.api.domain.ingredient.dto.response.IngredientResponse;
 
 public interface MemberIngredientCustomRepository {
     Slice<IngredientResponse> findMemberIngredientByCondition(MemberIngredientSearchCondition condition, Pageable pageable);
