+++ src/main/java/refooding/api/domain/ingredient/repository/MemberIngredientRepository.java
@@ -1,8 +1,8 @@
-package refooding.api.domain.fridge.repository;
+package refooding.api.domain.ingredient.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import refooding.api.domain.fridge.entity.MemberIngredient;
+import refooding.api.domain.ingredient.entity.MemberIngredient;
 
 import java.util.Optional;
